requirements.txt
Django==3.2.4
psycopg2-binary==2.9.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-import-export==2.7.0
djangorestframework==3.12.4